Extract LSPClient
